1.0.4
-silent mode
-referral system
-redesign interface

+++ users/cmd_usr_cash.py
@@ -51,14 +51,14 @@ async def kb(url):
 
 async def to_cancel(message,state:FSMContext):
 	res = await usr_info(message.from_user.id)
-	await bot.send_message(message.from_user.id,f'{await profile_info(message.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',reply_markup = await kb_main(res[2]))
+	await bot.send_message(message.from_user.id,f'{await profile_info(res[0],res[1],res[2])}',reply_markup = await kb_main(res[1]))
 	await bot.delete_message(message.from_user.id,message.message_id)
 	await state.finish()
 
 
 async def to_cancel_cb(call,state:FSMContext):
 	res = await usr_info(call.from_user.id)
-	await bot.edit_message_text(f'{await profile_info(call.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',call.from_user.id,call.message.message_id,reply_markup = await kb_main(res[2]))
+	await bot.edit_message_text(f'{await profile_info(res[0],res[1],res[2])}',call.from_user.id,call.message.message_id,reply_markup = await kb_main(res[1]))
 	await state.finish()
 
 	
@@ -75,7 +75,7 @@ async def deposit_token(call,callback_data:dict,state:FSMContext):
 	if value == 'back':
 		
 		res = await usr_info(call.from_user.id)
-		await bot.edit_message_text(f'{await profile_info(call.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',call.from_user.id,call.message.message_id,reply_markup = await kb_main(res[2]))
+		await bot.edit_message_text(f'{await profile_info(res[0],res[1],res[2])}',call.from_user.id,call.message.message_id,reply_markup = await kb_main(res[1]))
 		await state.finish()
 
 	else:
@@ -106,7 +106,7 @@ async def deposit_amount(message,state:FSMContext):
 	await set_deposit(message.from_user.id,inv_id,token_name,amount,message.date,'pending')
 
 	id_del = await bot.send_message(message.from_user.id,f'<b>ID</b> <code>{inv_id}</code>')
-	await bot.send_message(message.from_user.id,f'вы собираетесь внести депозит на сумму {amount} {token_name}\nсохраните id транзакции\n через 30 сек вы будете автоматически перенаправлены в главное меню',reply_markup=await kb(ext_link))
+	await bot.send_message(message.from_user.id,f'вы собираетесь внести депозит на сумму {amount} {token_name}\nсохраните id транзакции\n<i>скоро вы будете автоматически перенаправлены в главное меню</i>',reply_markup=await kb(ext_link))
 
 	await asyncio.sleep(30)
 
@@ -129,7 +129,7 @@ async def deposit_amount(message,state:FSMContext):
 		await set_deposit_status(message.from_user.id,inv_id,'not_paid')
 
 	res = await usr_info(message.from_user.id)
-	await bot.edit_message_text(f'{await profile_info(message.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',message.from_user.id,message.message_id+2,reply_markup = await kb_main(res[2]))
+	await bot.edit_message_text(f'{await profile_info(res[0],res[1],res[2])}',message.from_user.id,message.message_id+2,reply_markup = await kb_main(res[1]))
 	await state.finish()
 
 
@@ -149,7 +149,7 @@ async def withdraw_token(call,callback_data:dict,state:FSMContext):
 	if value == 'back':
 		
 		res = await usr_info(call.from_user.id)
-		await bot.edit_message_text(f'{await profile_info(call.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',call.from_user.id,call.message.message_id,reply_markup = await kb_main(res[2]))
+		await bot.edit_message_text(f'{await profile_info(res[0],res[1],res[2])}',call.from_user.id,call.message.message_id,reply_markup = await kb_main(res[1]))
 		await state.finish()
 
 	else:
@@ -164,7 +164,7 @@ async def withdraw_token(call,callback_data:dict,state:FSMContext):
 
 
 				res1 = await usr_info(call.from_user.id)
-				await bot.edit_message_text(f'{await profile_info(call.from_user.id,res1[0],res1[1],res1[2],res1[3],res1[4],res1[5])}',call.from_user.id,call.message.message_id,reply_markup = await kb_main(res1[2]))
+				await bot.edit_message_text(f'{await profile_info(res1[0],res1[1],res1[2])}',call.from_user.id,call.message.message_id,reply_markup = await kb_main(res1[1]))
 				await state.finish()
 
 			elif float(res)!=0.0 and float(res) >= float(min_with):
@@ -209,7 +209,7 @@ async def withdraw_amount(message,state:FSMContext):
 
 				amount = min_amnt_with
 
-			if vip=='0':
+			if vip=='💤':
 
 				await state.finish()
 				await bot.send_message(user_id,f'id <code>{trs_id}</code>: Ваша заявка на вывод средств принята в обработку.\n<i>сохраните id транзакции</i>')
@@ -220,7 +220,7 @@ async def withdraw_amount(message,state:FSMContext):
 				
 				await asyncio.sleep(5)
 				res = await usr_info(message.from_user.id)
-				await bot.send_message(message.from_user.id,f'{await profile_info(message.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',reply_markup = await kb_main(res[2]))
+				await bot.send_message(message.from_user.id,f'{await profile_info(res[0],res[1],res[2])}',reply_markup = await kb_main(res[1]))
 
 				
 			else:
@@ -231,7 +231,7 @@ async def withdraw_amount(message,state:FSMContext):
 
 				await asyncio.sleep(5)
 				res = await usr_info(message.from_user.id)
-				await bot.send_message(message.from_user.id,f'{await profile_info(message.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',reply_markup = await kb_main(res[2]))
+				await bot.send_message(message.from_user.id,f'{await profile_info(message.from_user.id,res[0],res[1],res[2])}',reply_markup = await kb_main(res[1]))
 				await state.finish()
 			
 
@@ -241,7 +241,7 @@ async def withdraw_amount(message,state:FSMContext):
 			await message.answer('вронг дата. телл зе админ абоут ит')
 			await bot.send_message(ADMIN, 'some happened wrong. user balance equal as main balance')
 			res = await usr_info(message.from_user.id)
-			await bot.send_message(message.from_user.id,f'{await profile_info(message.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',reply_markup = await kb_main(res[2]))
+			await bot.send_message(message.from_user.id,f'{await profile_info(res[0],res[1],res[2])}',reply_markup = await kb_main(res[1]))
 			await state.finish()
 
 			
@@ -249,7 +249,7 @@ async def withdraw_amount(message,state:FSMContext):
 			await message.answer('вронг дата. телл зе админ абоут ит')
 			await bot.send_message(ADMIN, ' big amount of withdraw.some happened wrong')
 			res = await usr_info(message.from_user.id)
-			await bot.send_message(message.from_user.id,f'{await profile_info(message.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',reply_markup = await kb_main(res[2]))
+			await bot.send_message(message.from_user.id,f'{await profile_info(res[0],res[1],res[2])}',reply_markup = await kb_main(res[1]))
 			await state.finish()
 
 
@@ -261,11 +261,13 @@ async def withdraw_amount(message,state:FSMContext):
 
 
 async def spam_remover(message,state:FSMContext):
-	if message.chat.type != 'supergroup':
-		if message.text.lower() == 'reload':
-			res = await usr_info(message.from_user.id)
-			await bot.send_message(message.from_user.id,f'{await profile_info(message.from_user.id,res[0],res[1],res[2],res[3],res[4],res[5])}',reply_markup = await kb_main(res[2]))
-			await state.finish()
-			
-		elif message.text.lower()!='cancel' or float(message.text.isdigit())!=True:
-			await bot.delete_message(message.from_user.id,message.message_id)
+	if str(message.chat.id) in ADMIN:
+		print('spam checker')
+		if message.chat.type != 'supergroup':
+			if message.text.lower() == 'reload':
+				res = await usr_info(message.from_user.id)
+				await bot.send_message(message.from_user.id,f'{await profile_info(res[0],res[1],res[2])}',reply_markup = await kb_main(res[1]))
+				await state.finish()
+
+			elif message.text.lower()!='cancel' or float(message.text.isdigit())!=True:
+				await bot.delete_message(message.from_user.id,message.message_id)
